+++ CHANGELOG.md
@@ -15,6 +15,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Reset to defaults works with settings as well
 
+### Fixed
+
+- Creating metadata of a resource definition that has multiple parents
+
 ## [1.11.0]
 
 ### Added